Massive lint fixes incoming.
