+++ src/api/BaseApi.config.js
@@ -1,11 +1,11 @@
 export const config = () => {
   if (__DEV__) {
-    return {'apiRoot': '', version: 'v3' }
+    return { 'apiRoot': '', version: 'v3' }
   }
 
   if (__TEST__) {
-    return {'apiRoot': '/', version: 'v3' }
+    return { 'apiRoot': '/', version: 'v3' }
   }
 
-  return {'apiRoot': '/', version: 'v3' }
+  return { 'apiRoot': '/', version: 'v3' }
 }
