+++ src/ui/@state/@region/Region.selectors.js
@@ -293,9 +293,7 @@ export const getCountyListSelector = createSelector(
     }
 
     const countiesInSelectedRegionDictionary = regionIndex[regionName]
-    console.log(countiesInSelectedRegionDictionary)
     const countyIdsInSelectedRegion = keys(countiesInSelectedRegionDictionary)
-    console.log(countyIdsInSelectedRegion)
     const countyObjects = countyIdsInSelectedRegion.map((id) => {
       const streamIdsInRegion = countiesInSelectedRegionDictionary[id]
       const county = countiesDictionary[id]
@@ -332,7 +330,6 @@ export const getCountyListSelector = createSelector(
         console.log(visibleStreamsInCounty) // eslint-disable-line
       }
     })
-    console.log(countyObjects)
     const filteredCountyObjects = countyObjects
       .sort((a,b) => a.name.localeCompare(b.name))
       .filter((x) => x.streams.length > 0)
