+++ src/ui/@state/@region/map/MapboxGlMap/styles/Base.style.js
@@ -2113,19 +2113,6 @@ const BaseStyle = {
             ],
           ],
         },
-        'line-opacity': {
-          'base': 1,
-          'stops': [
-            [
-              13.9,
-              0,
-            ],
-            [
-              14,
-              1,
-            ],
-          ],
-        },
       },
     },
     {
