+++ src/ui/core/CoreLayout.container.js
@@ -31,11 +31,6 @@ class CoreLayoutContainer extends Component {
 
   listenToRoutes () {
     const {router} = this.props
-    if (router == null) {
-      console.log('No router found. Check Minimap component')
-      return
-    }
-
     router.listen(({pathname}) => {
       AnonymousAnalyzerApi.recordEvent('page_navigation', {})
     })
