+++ src/ui/core/header/minimap/svgMinimap/AdministrativeLayer.component.js
@@ -84,7 +84,7 @@ AdministrativeLayerComponent.propTypes = {
   'cachedRegions': PropTypes.object.isRequired,
   'isOffline': PropTypes.bool.isRequired,
   pathGenerator: PropTypes.func.isRequired,
-  selectRegion: PropTypes.object.isRequired,
+  selectRegion: PropTypes.func.isRequired,
 }
 
 export default AdministrativeLayerComponent
