+++ src/ui/core/header/subtitle/Subtitle.selectors.js
@@ -20,7 +20,6 @@ export const subtitleSelector = createSelector(
       return WELCOME_TITLE
     }
 
-    console.log('hello', selectedRegion)
     const isOnlyStateSelected = selectedState != null && selectedRegion == null
     if (isOnlyStateSelected) {
       const state = selectedState.properties.name
