+++ src/ui/core/offline/Offline.selectors.js
@@ -1,6 +1,6 @@
 import {isEmpty, keyBy} from 'lodash'
 import {createSelector} from 'reselect'
-import {regionsDictionarySelector, selectedRegionPathKeySelector} from 'ui/core/Core.selectors'
+import {regionsDictionarySelector} from 'ui/core/Core.selectors'
 export const isOfflineSelector = (state) => {
   if (state == null) {
     return false
@@ -44,27 +44,27 @@ export const cachedRegionsDictionary = createSelector(
     const cachedRegionDictionary = endpoints
       .filter(x => x.indexOf('v3') >= 0)
       .reduce((dictionary, endpoint) => {
-      const tokens = endpoint.split('/')
-        .filter((x) => x.length > 0)
-      const isNotStateRegion = tokens.length <= 3
-      if (isNotStateRegion) {
-        return dictionary
-      }
+        const tokens = endpoint.split('/')
+          .filter((x) => x.length > 0)
+        const isNotStateRegion = tokens.length <= 3
+        if (isNotStateRegion) {
+          return dictionary
+        }
 
-      const regionFileName = tokens[3]
-      const stateName = tokens[2]
-      const isTopojsonFile = regionFileName.indexOf('.topo.json') >= 0
-      if (isTopojsonFile === false) {
+        const regionFileName = tokens[3]
+        const stateName = tokens[2]
+        const isTopojsonFile = regionFileName.indexOf('.topo.json') >= 0
+        if (isTopojsonFile === false) {
+          return dictionary
+        }
+        const regionName = regionFileName.split('.')[0]
+        // I believe we index by name.
+        const regionKey = `${stateName}/${regionName}`
+        const region = regionDictionary[regionKey]
+        const regionId = region.properties.gid
+        dictionary[regionId] = region
         return dictionary
-      }
-      const regionName = regionFileName.split('.')[0]
-      // I believe we index by name.
-      const regionKey = `${stateName}/${regionName}`
-      const region = regionDictionary[regionKey]
-      const regionId = region.properties.gid
-      dictionary[regionId] = region
-      return dictionary
-    }, {})
+      }, {})
 
     return cachedRegionDictionary
   })
