+++ src/ui/core/termsOfAgreement/Agreement.stories.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import { storiesOf, action, linkTo } from '@kadira/storybook'
+import { storiesOf } from '@kadira/storybook'
 import AgreementComponent from './Agreement.component'
 
 storiesOf('core/termsOfAgreement', module)
