+++ src/ui/core/termsOfAgreement/PrivacyPolicy.component.js
@@ -1,3 +1,4 @@
+/* eslint-disable max-lines */
 import React, {Component} from 'react'
 import PropTypes from 'prop-types'
 import classes from './Agreement.scss'
