+++ src/ui/core/termsOfAgreement/TermsOfService.stories.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import { storiesOf, action, linkTo } from '@kadira/storybook'
+import { storiesOf } from '@kadira/storybook'
 import TermsOfService from './TermsOfService.component'
 
 storiesOf('core/termsOfAgreement', module)
