+++ src/ui/core/termsOfAgreement/ThankYou.stories.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import { storiesOf, action, linkTo } from '@kadira/storybook'
+import { storiesOf } from '@kadira/storybook'
 import ThankYou from './ThankYou.component'
 
 storiesOf('core/termsOfAgreement', module)
